Bugfix/lattice 2017 delete edges locking

+++ src/main/java/com/openlattice/datastore/edm/controllers/EdmController.java
@@ -486,11 +486,15 @@ public class EdmController implements EdmApi, AuthorizingComponent, AuditingComp
             throw new ForbiddenException( "You shall not pass!" );
         }
 
+        // linking entitysets have no entities or associations
+        if(!entitySet.isLinking()) {
+            // associations need to be deleted first, because edges are deleted in DataGraphManager.deleteEntitySet call
+            deleteAssociationsOfEntitySet( entitySetId );
+            dgm.deleteEntitySet( entitySetId, authorizedPropertyTypes );
+        }
+
         modelService.deleteEntitySet( entitySetId );
         securableObjectTypes.deleteSecurableObjectType( new AclKey( entitySetId ) );
-        // associations need to be deleted first, because edges are deleted in DataGraphManager.deleteEntitySet call
-        deleteAssociationsOfEntitySet( entitySetId );
-        dgm.deleteEntitySet( entitySetId, authorizedPropertyTypes );
 
         recordEvent( new AuditableEvent(
                 getCurrentUserId(),
